refactor: moved note creation functionality to the component
